adds some more network stuff

+++ lib/adaptibrew.rb
@@ -1,11 +1,15 @@
 require 'git'
 require 'fileutils'
+require_relative 'helpers'
+
+include Helpers
 
 # This is the 'manager' for the adaptibrew repo. It handles cloning and such.
 class Adaptibrew
 
   # If used in IRB, Ripl, etc. it will clone into the directory IRB was started in
   def clone(path=nil)
+    raise "🛑  Cannot clone, no network connection" unless network?
     if !Dir.exists?('adaptibrew')
       Git.clone('https://github.com/adaptiman/adaptibrew.git', 'adaptibrew', :path => path)
     end
@@ -14,11 +18,17 @@ class Adaptibrew
 
   # Danger zone...
   def clear()
-    FileUtils.rm_rf('adaptibrew')
+    if !network?
+      print "Warning: you have no network connection. If you clear, you will not be able to clone again, and you'll be stuck without the adaptibrew source. Are you sure? "
+      if confirm
+        FileUtils.rm_rf('adaptibrew')
+      end
+    end
     self
   end
 
   def refresh
+    raise "🛑  Cannot refresh, no network connection" unless network?
     clear
     clone
     self
